validation touch up
